requirements.txt
apache-airflow==2.7.0
pytest==7.4.3
pytest-cov==4.1.0